allow option and target specification for `//tools:lint`
Allow arguments to be specified for the `@rules_clang_tidy` aspect, as
well as target specification.

If no arguments are passed to `//tools:lint`, linting applies to all
targets in the workspace using a wildcard label. If any argument is
passed, even a non-label, the lint targets must be explicitly
specified.

Change-Id: Id0f236dd9929deccd495d674121daf64f1223859