refactor `blade_type_from(_dimension)` traits (#86)
Define trait `blade_type_from_dimension`, which determines the `blade`
type given the algebra type and a random access range as a non-type
template
parameter. The random access range specifies the dimensions of the
`blade`.

Update the `blade_type_from` trait to take an algebra type and a
non-type template parameter that is either a mask or a range of
dimensions. A range of `bool` is interpreted as a mask. A range of
`std::size_t` is interpreted as dimensions.

Change-Id: I08d04592475f294e755e5fef88d1dfd7cbfdaea6