fix blade order
`blade`s are stored in a `multivector` based on a total order. This
commit updates the `blade` total order to match the blade order of lines
and planes in Lengyel's notation.

This order is well defined up 4 dimension. For blades with basis
dimensions `i...` and `j...`:
- if sizeof...(i) < sizeof...(j), then b_{i...} < b_{j...}
- otherwise if sizeof...(i) == sizeof...(j)
  and if i... contains 0 and j... does not contain 0, then b_{i...} < b_{j...}
- otherwise if i... and j... both contain 0 or both do not contain 0
  and if sizeof...(i) == 1 the ordering of b_{i...} <=> b_{j...} is determined
  by i...[0] <=> j...[0]
- otherwise if sizeof...(i) != 1 the ordering of b_{i...} <=> b_{j...} is
  determined by bitmask(j...).to_unsigned() <=> bitmask(i...).to_unsigned()

For G3 (dimension 4), this order results in the following layouts for
geometric types:
- point: [ weight | position ]
- line: [ direction | moment ]
- plane : [ normal | position ]

Its unclear how blade ordering impacts layout for dimensions
greater than 4.

Change-Id: If6aec4167cd5f6780c609cae333da7fad954d268