fix basis dimension order
Change blade basis dimension ordering. This new ordering is sometimes,
but not always, lexicographical. Using a different ordering reduces the
need to negate elements of line and plane types.

For grade 1 blades, canonical order is equivalent to lexicographic
order.

For grade 2 blades without a 0-dimensional basis (the projective
dimension), canonical order is equivalent to cyclic order:
* (1, 2) -> (1, 2)
* (1, 3) -> (3, 1)
* (2, 3) -> (2, 3)

For grade 2+ blades with a 0-dimensional basis (the projective
dimension), canonical order of [0, i...] is equivalent to
[0, canonical-order([i...])].

Change-Id: I44150916fdfcbbfdf2b7dac5088a6345046c391d